+++ src/EnumServiceProvider.php
@@ -4,6 +4,7 @@ namespace LeMaX10\Enums;
 
 use Illuminate\Support\ServiceProvider;
 use Illuminate\Support\Facades\Validator;
+use LeMaX10\Enums\Rules\EnumValue;
 
 /**
  * Class EnumServiceProvider
@@ -26,12 +27,11 @@ class EnumServiceProvider extends ServiceProvider
     {
         Validator::extend('enum', static function ($attribute, $value, $parameters, $validator) {
             $enum = $parameters[0] ?? null;
-            if (!class_exists($enum)) {
+            if (!$enum) {
                 return false;
             }
 
-            $enumValidator = call_user_func([$enum, 'rule']);
-            return $enumValidator->passes($attribute, $value);
+            return (new EnumValue($enum))->passes($attribute, $value);
         });
     }
 }