+++ CHANGELOG.md
@@ -11,3 +11,8 @@
 - Added Rule static helper for use in validation
 - Added additional check of string rule by enum
 - Fixed style of calling static methods in some package objects
+## 1.3.5 - 2020-05-03
+- Fix error equals enums in model
+## 1.4.0 - 2020-05-20
+- Add EnumList Trait
+- Fix Enum Value Validation